Bug fixes and prefab additions for v0.10 release

+++ Assets/Scripts/Fishables/Fishable.cs
@@ -11,7 +11,7 @@ namespace Fishing.Fishables
         [SerializeField] private string itemDescription;
         private float weight;
         private float length;
-        [SerializeField] private int baseValue;
+        [SerializeField] private float baseValue;
         private float actualValue;
 
         [Header("Variation")]
